+++ src/Model/Produto/Produto.pas
@@ -15,14 +15,11 @@ type
     FReferencia: String;
     FPreco: Double;
     FCategoria : String;
-    FQuantidadeEstoque: Integer;
+    FQuantidadeEstoque: Double;
     FFornecedor: String;
     FDataValidade: TDate;
 
   public
-//    constructor Create(const CID: Integer; CNome: String; CCodigoBarras: String;
-//     CDescricao: String; CReferencia: String; CPreco: Double; CFornecedor: String;
-//     CDataValidade: TDate; CCategoria: String; CQuantidadeEstoque: Integer) overload;
 
     property ID: Integer read FID write FID;
     property Nome: String read FNome write FNome;
@@ -31,7 +28,7 @@ type
     property Referencia: String read FReferencia write FReferencia;
     property Preco: Double read FPreco write FPreco;
     property Categoria: String read FCategoria write FCategoria;
-    property QuantidadeEstoque: Integer read FQuantidadeEstoque write FQuantidadeEstoque;
+    property QuantidadeEstoque: Double read FQuantidadeEstoque write FQuantidadeEstoque;
     property Fornecedor: String read FFornecedor write FFornecedor;
     property DataValidade: TDate read FDataValidade write FDataValidade;
